requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
bs4
requests
youtube-transcript-api
pytube
langchain-text-splitters
numexpr
gradio
duckduckgo-search
ddgs
pypdf
sentence-transformers
torch
chromadb
pandas